server/requirements.txt
fastapi>=0.109
uvicorn[standard]>=0.27
websockets>=12.0
python-dotenv>=1.0
requests>=2.31
aiohttp>=3.9
PyJWT>=2.8
twilio>=9.0
openai>=1.3
edge-tts>=7.0
numpy>=1.26
python-dateutil>=2.8
PyAudio>=0.2.14
